+++ lib/AppInfo/Application.php
@@ -14,6 +14,7 @@ use OCP\AppFramework\Bootstrap\IRegistrationContext;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 
+use OCA\SuiteCRM\Dashboard\SuiteCRMCalendarWidget;
 use OCA\SuiteCRM\Dashboard\SuiteCRMWidget;
 use OCA\SuiteCRM\Search\SuiteCRMSearchProvider;
 use OCA\SuiteCRM\Notification\Notifier;
@@ -33,6 +34,7 @@ class Application extends App implements IBootstrap {
 
 	public function register(IRegistrationContext $context): void {
 		$context->registerDashboardWidget(SuiteCRMWidget::class);
+		$context->registerDashboardWidget(SuiteCRMCalendarWidget::class);
 		$context->registerSearchProvider(SuiteCRMSearchProvider::class);
 		$context->registerNotifierService(Notifier::class);
 	}